- break lua config into modular files - update .luacheckrc to use full lua stdlib - notify on buffer releoad
