+++ bash/posix.sh
@@ -30,25 +30,8 @@ posix_git() {
 }
 
 posix_nvim() {
-  rm -rf $HOME/.config/nvim
-  mkdir -p $HOME/.config/nvim
-  ln -s $PWD/init.lua $HOME/.config/nvim/init.lua
-  msg "${OK}${FUNCNAME[0]}: symlinked nvim lua config."
-
-  # Clean up all plugins including paq
-  NVIM_PLUGIN_PATH="$HOME/.local/share/nvim/site/pack"
-  rm -rf "$NVIM_PLUGIN_PATH"
-  mkdir -p "$NVIM_PLUGIN_PATH/paqs/start"
-  msg "${OK}${FUNCNAME[0]}: cleaned up all nvim plugins."
-
-  # Install paq-nvim
-  PAQ_PATH="$HOME/.local/share/nvim/site/pack/paqs/start/paq-nvim"
-  git clone --depth=1 https://github.com/savq/paq-nvim "$PAQ_PATH" > /dev/null 2>&1
-  msg "${OK}${FUNCNAME[0]}: installed paq-nvim."
-
-  # Install plugins
-  nvim --headless -c 'autocmd User PaqDoneInstall quit' -c 'PaqInstall' > /dev/null 2>&1
-  msg "${OK}${FUNCNAME[0]}: installed neovim plugins."
+  _posix_nvim_config "${FUNCNAME[0]}"
+  _posix_nvim_plugins "${FUNCNAME[0]}"
 }
 
 posix_symlinks() {
@@ -66,4 +49,54 @@ main_posix() {
   posix_git
   posix_nvim
   posix_asdf
+}
+
+# Private functions
+
+_posix_nvim_config() {
+  local caller="$1"
+  
+  # Ensure nvim config directory exists
+  rm -rf $HOME/.config/nvim
+  mkdir -p $HOME/.config/nvim
+
+  # Create a loader init.lua that points to the dotfiles repo
+  cat > $HOME/.config/nvim/init.lua << EOF
+-- This file is auto-generated by dotfiles setup
+-- It loads the actual configuration from the dotfiles repository
+
+-- Check if we're in installation mode or normal operation
+if vim.env.NVIM_INSTALL_MODE == "1" then
+  -- In installation mode, load the plugins registry directly to get the plugin definitions
+  dofile("$PWD/neovim/plugins/init.lua")
+  
+  -- Then run the installation
+  vim.cmd("autocmd User PaqDoneInstall quit")
+  vim.cmd("PaqInstall")
+else
+  -- In normal operation, load the main neovim config
+  dofile("$PWD/neovim.lua")
+end
+EOF
+
+  msg "${OK}${caller}: created nvim config that loads from dotfiles repo."
+}
+
+_posix_nvim_plugins() {
+  local caller="$1"
+  
+  # Clean up all plugins including paq
+  NVIM_PLUGIN_PATH="$HOME/.local/share/nvim/site/pack"
+  rm -rf "$NVIM_PLUGIN_PATH"
+  mkdir -p "$NVIM_PLUGIN_PATH/paqs/start"
+  msg "${OK}${caller}: cleaned up all nvim plugins."
+
+  # Install paq-nvim
+  PAQ_PATH="$HOME/.local/share/nvim/site/pack/paqs/start/paq-nvim"
+  git clone --depth=1 https://github.com/savq/paq-nvim "$PAQ_PATH" > /dev/null 2>&1
+  msg "${OK}${caller}: installed paq-nvim."
+
+  # Launch Neovim to install plugins - redirect both stdout and stderr
+  NVIM_INSTALL_MODE=1 nvim --headless -c 'autocmd User PaqDoneInstall quit' > /dev/null 2>&1
+  msg "${OK}${caller}: installed neovim plugins."
 }
