+++ neovim/plugins/diffview.lua
@@ -1,0 +1,8 @@
+-- Diffview plugin configuration
+
+-- Behavior
+
+-- Display
+
+-- Keybindings
+vim.keymap.set('n', '<Leader>d', ':execute \'DiffviewOpen\'<CR>', { silent = true })
