+++ neovim/plugins/marv.lua
@@ -1,0 +1,8 @@
+-- Marv plugin configuration
+
+-- Behavior
+
+-- Display
+
+-- Keybindings
+vim.keymap.set('n', '<Leader>m', ':execute \'MarvToggle\'<CR>', { silent = true })
