+++ neovim/plugins/nerdtree.lua
@@ -1,0 +1,9 @@
+-- NERDTree plugin configuration
+
+-- Behavior
+vim.g.NERDTreeShowHidden = 1
+
+-- Display
+
+-- Keybindings
+vim.keymap.set('n', '<Leader>n', ':NERDTreeToggle<CR>', { silent = true })
