+++ neovim/plugins/rainbow_parentheses.lua
@@ -1,0 +1,13 @@
+-- Rainbow Parentheses plugin configuration
+
+-- Behavior
+vim.cmd [[
+  augroup rainbow_parentheses
+    autocmd!
+    autocmd BufEnter * RainbowParentheses
+  augroup END
+]]
+
+-- Display
+
+-- Keybindings
