+++ neovim/plugins/vim_bufkill.lua
@@ -1,0 +1,8 @@
+-- Vim-Bufkill plugin configuration
+
+-- Behavior
+
+-- Display
+
+-- Keybindings
+vim.keymap.set('n', '<C-b>q', ':BD!<CR>', { silent = true })
