+++ neovim/plugins/vim_maximizer.lua
@@ -1,0 +1,8 @@
+-- Vim-Maximizer plugin configuration
+
+-- Behavior
+
+-- Display
+
+-- Keybindings
+vim.keymap.set('n', '<C-w>z', ':MaximizerToggle<CR>', { silent = true })
