+++ neovim/plugins/winresizer.lua
@@ -1,0 +1,8 @@
+-- Winresizer plugin configuration
+
+-- Behavior
+vim.g.winresizer_start_key = '<silent><C-w>r'
+
+-- Display
+
+-- Keybindings
