+++ neovim/plugins/wisp.lua
@@ -1,0 +1,8 @@
+-- Wisp (Syfe) plugin configuration
+
+-- Behavior
+
+-- Display
+
+-- Keybindings
+vim.keymap.set('n', '<Leader>w', ':execute \'Wisp\'<CR>', { silent = true })
